Refactor tag command to use a proper tiered-architecture structure
 - we've managed to get back our inversion of control.
 - we need to go back and refactor the pic command to follow this format too
 - need to think about renaming the commands module and double check proper rust naming conventions
