+++ src/commands/tag.rs
@@ -1,28 +1,9 @@
+use crate::data_access::tag_repository::TagRepository;
 use crate::discord_client::CommandHandler;
-use rusqlite::Connection;
-use rusqlite::NO_PARAMS;
-
-trait TagRepository {
-    fn create_tag(
-        &self,
-        tag_name: &str,
-        tag_body: &str,
-        tennant_id: u64,
-    ) -> Result<(), Box<std::error::Error>>;
-    fn read_tag(&self, tag_name: &str, tennant_id: u64) -> Result<String, Box<std::error::Error>>;
-    fn read_all_tags(&self, tennant_id: u64) -> Result<Vec<Tag>, Box<std::error::Error>>;
-}
+use crate::models::tag::Tag;
 
 pub struct TagCommand {
     tag_repository: Box<TagRepository + Send>,
-}
-pub struct SqliteTagRepository {
-    db_connection: Connection,
-}
-
-struct Tag {
-    name: String,
-    body: String,
 }
 
 fn parse_ntag(command: &str) -> Option<Tag> {
@@ -51,9 +32,9 @@ fn parse_tag_command(command: &str) -> Option<&str> {
 }
 
 impl TagCommand {
-    pub fn new() -> Result<TagCommand, Box<std::error::Error>> {
-        let tag_repository = Box::new(SqliteTagRepository::new()?);
-
+    pub fn new(
+        tag_repository: Box<TagRepository + Send>,
+    ) -> Result<TagCommand, Box<std::error::Error>> {
         Ok(TagCommand { tag_repository })
     }
 }
@@ -105,79 +86,6 @@ impl CommandHandler for TagCommand {
         }
 
         Ok(())
-    }
-}
-
-impl SqliteTagRepository {
-    fn new() -> Result<SqliteTagRepository, Box<std::error::Error>> {
-        let db_connection = Connection::open("tags.db")?;
-
-        db_connection.execute(
-            "CREATE TABLE IF NOT EXISTS tags (
-            id INTEGER PRIMARY KEY,
-            name TEXT NOT NULL,
-            body TEXT NOT NULL,
-            tennant_id INTEGER NOT NULL,
-            UNIQUE (name, tennant_id)
-        )",
-            NO_PARAMS,
-        )?;
-
-        Ok(SqliteTagRepository { db_connection })
-    }
-}
-
-impl TagRepository for SqliteTagRepository {
-    fn create_tag(
-        &self,
-        tag_name: &str,
-        tag_body: &str,
-        tennant_id: u64,
-    ) -> Result<(), Box<std::error::Error>> {
-        //sqlite3 doesn't support u64 properly so we have to cast to i64 first. as long as we do this consistantly we shouldn't have problems.
-        let tennant_id = (tennant_id as i64).to_string();
-
-        self.db_connection.execute(
-            "INSERT INTO tags
-            (name, body, tennant_id) VALUES (?1, ?2, ?3)",
-            &[tag_name, tag_body, &tennant_id],
-        )?;
-        Ok(())
-    }
-
-    fn read_tag(&self, tag_name: &str, tennant_id: u64) -> Result<String, Box<std::error::Error>> {
-        //sqlite3 doesn't support u64 properly so we have to cast to i64 first. as long as we do this consistantly we shouldn't have problems.
-        let tennant_id = (tennant_id as i64).to_string();
-
-        Ok(self
-            .db_connection
-            .prepare(
-                "SELECT body
-                FROM tags 
-                WHERE tennant_id = (?1) AND name = (?2)",
-            )?
-            .query_row(&[&tennant_id, tag_name], |row| Ok(row.get(0)?))?)
-    }
-
-    fn read_all_tags(&self, tennant_id: u64) -> Result<Vec<Tag>, Box<std::error::Error>> {
-        //sqlite3 doesn't support u64 properly so we have to cast to i64 first. as long as we do this consistantly we shouldn't have problems.
-        let tennant_id = (tennant_id as i64).to_string();
-
-        Ok(self
-            .db_connection
-            .prepare(
-                "SELECT name, body
-                FROM tags
-                WHERE tennant_id = (?1)",
-            )?
-            .query_map(&[&tennant_id], |row| {
-                Ok(Tag {
-                    name: row.get(0)?,
-                    body: row.get(1)?,
-                })
-            })?
-            .filter_map(Result::ok)
-            .collect())
     }
 }
