+++ src/lib.rs
@@ -1,2 +1,4 @@
 pub mod commands;
+pub mod data_access;
 pub mod discord_client;
+pub mod models;
