+++ src/models/tag.rs
@@ -1,0 +1,4 @@
+pub struct Tag {
+    pub name: String,
+    pub body: String,
+}
